requirements.txt
# Traitement d'image et manipulation
Pillow

# Traitement scientifique (pour manipulations avancées, matrices, etc.)
numpy

# Segmentation, détection de pose, estimation profondeur
opencv-python
mediapipe
# torch et torchvision installés à part, version CUDA uniquement

# Pour MiDaS (estimation de profondeur)
timm==0.6.13

# Stable Diffusion et ControlNet (inpainting, génération, etc.)
diffusers[torch]
transformers
# Pour ControlNet/Stable Diffusion, gestion de modèles, prompts, etc.
accelerate
safetensors

# Utilitaires divers
tqdm

# Pour l'inpainting avancé, optionnel
scipy

onnxruntime